+++ packages/components/src/card/card/hook.js
@@ -1,12 +1,4 @@
 /**
- * External dependencies
- */
-// Disable reason: Temporarily disable for existing usages
-// until we remove them as part of https://github.com/WordPress/gutenberg/issues/30503#deprecating-emotion-css
-// eslint-disable-next-line no-restricted-imports
-import { cx } from '@emotion/css';
-
-/**
  * WordPress dependencies
  */
 import deprecated from '@wordpress/deprecated';
@@ -18,6 +10,7 @@ import { useMemo } from '@wordpress/element';
 import { useContextSystem } from '../../ui/context';
 import { useSurface } from '../../surface';
 import * as styles from '../styles';
+import { useCx } from '../../utils/hooks/use-cx';
 
 /**
  * @param {import('../../ui/context').PolymorphicComponentProps<import('../types').Props, 'div'>} props
@@ -58,6 +51,8 @@ export function useCard( props ) {
 		size = 'medium',
 		...otherProps
 	} = useContextSystem( useDeprecatedProps( props ), 'Card' );
+
+	const cx = useCx();
 
 	const classes = useMemo( () => {
 		return cx(